Daemon: One session manages two flows
IN_PORT will be the one to insert offload flows for both IN and OUT
ports. Aging will be only done when the flow_in is aging out, and
both flow_in and flow_out will be removed at the same time.

Flow query now return the out_packets which queries flow_out.

Signed-off-by: Bodong Wang <bodong@mellanox.com>