Daemon: Workaround for no update on counter of one direction
Each flow in one session has a ref count, it will be set as 1 when
added, set as 0 when aged out. Only when ref counts on both flows
are 0, session is deleted.

Note that dpdk aging seems using age context as the matching criteria
to look up flow, as if two flows has the same context, only one of
them get aging out.

Fixes: 6549eaa2c4a1 ("Daemon: Only delete session when both directions are aged out")
Signed-off-by: Bodong Wang <bodong@mellanox.com>